% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{fINDexers}
\alias{fINDexers}
\alias{\%ci\%}
\alias{\%ri\%}
\alias{\%mi\%}
\alias{\%ai\%}
\title{Find row or column by name or index}
\usage{
c \%ci\% d

r \%ri\% d

rc \%mi\% d

nv \%ai\% d
}
\arguments{
\item{c}{Column name or index}

\item{d}{A named vector, list, matrix, or data frame}

\item{r}{Row name or index}

\item{rc}{A 2-element numeric or character vector representing \code{c(r,c)}. Names (character) and indices (numeric) vectors can be mixed if \code{rc} is passed as a 2-element list object.}

\item{nv}{A numeric value, or vector of values of which you want to know the indices in \code{d}.}
}
\value{
If \code{r/c/rc} is numeric, the name corresponding to the row/column index of \code{d}, if \code{r/c/rc} is a character vector, the row/column index corresponding to the row/column name. If \code{dimnames(d) == NULL}, but \code{names(d) != NULL} then \code{\%ci\%} and \code{\%ri\%} will look up \code{r/c} in \code{names(d)}
}
\description{
Find row or column by name or index

Column by name or index

Row by name or number

Matrix cell index by name or number

Return all indices of a (range of) values
}
\examples{



}
\author{
Fred Hasselman
}
