% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{\%]\%}
\alias{\%]\%}
\title{Extract vector rear}
\usage{
i \%]\% x
}
\arguments{
\item{i}{An index in the range \code{[1,length(x)]}}

\item{x}{A vector}
}
\value{
The contents of vector \code{x} in the range \code{x[i:length(x)]}
}
\description{
Extract vector rear
}
