% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{\%e\%}
\alias{\%e\%}
\title{Is element of... with multiple input types}
\usage{
x \%e\% y
}
\arguments{
\item{x}{A vector, data frame of list containing numbers and/or characters that could be elements of y}

\item{y}{An object that could contain values in x}
}
\value{
Logical vector indicating which x are an element of y
}
\description{
Is element of... with multiple input types
}
