% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{\%++\%}
\alias{\%++\%}
\title{Non-negative increment}
\usage{
counter \%++\% increment
}
\arguments{
\item{counter}{If \code{counter} \eqn{\ge 0} and \code{increment} \eqn{> 0} and are both integers, \code{counter} will change by the value of \code{increment}.}

\item{increment}{An integer value \eqn{> 0} to add to \code{counter}}
}
\description{
Increment a counter by an arbitrary interval greater than 0.
}
\examples{

# Notice the difference between passing an object and a value for counter

# Value
(0 \%++\% 5)
(0 \%++\% 5)

# Object
i <- 0
(i \%+-\% 5)
(i \%+-\% 5)

# This means we can use the infix in a while ... statement
# WARNING: As is the case for any while ... statement, be careful not to create an infinite loop!

while(i < 20) i \%+-\% 5

}
\seealso{
%+-%

Other auto counters: \code{\link{\%+-\%}}
}
\author{
Fred Hasselman
}
