% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{insiders}
\alias{insiders}
\alias{\%[]\%}
\alias{\%()\%}
\alias{\%[)\%}
\alias{\%(]\%}
\alias{\%[.]\%}
\alias{\%(.)\%}
\alias{\%[.)\%}
\alias{\%(.]\%}
\title{Values inside interval}
\usage{
x \%[]\% j

x \%()\% j

x \%[)\% j

x \%(]\% j

x \%[.]\% j

x \%(.)\% j

x \%[.)\% j

x \%(.]\% j
}
\arguments{
\item{x}{A vector}

\item{j}{A 2-element numeric vector indicating a range}
}
\value{
Logical vector of length \code{x}, or, values in the range \code{j}
}
\description{
Values inside interval

In closed interval: \verb{\%[]\%}

In open interval: \verb{\%()\%}

In half-closed interval (left): \verb{\%[)\%}

In half-closed interval (right): \verb{\%(]\%}

Return x in closed interval: \verb{\%[.]\%}

Return x in open interval: \verb{\%(.)\%}

Return x in half-closed interval (left): \verb{\%[.)\%}

Return x in half-closed interval (right): \verb{\%(.]\%}
}
\note{
Package \code{DescTools} provides similar functions
}
\examples{

# Closed interval
0:5 \%[]\% c(1,5)  # logical vector
0:5 \%[.]\% c(1,5) # extract values

# Open interval
0:5 \%()\% c(1,5)
0:5 \%(.)\% c(1,5)

# Closed interval left
0:5 \%[)\% c(1,5)
0:5 \%[.)\% c(1,5)

# Closed interval right
0:5 \%(]\% c(1,5)
0:5 \%(.]\% c(1,5)


}
