% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Wholenumber check}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Number to check}

\item{tol}{Tolerance}
}
\value{
TRUE if \code{x} is a signed integer.
}
\description{
Wholenumber check
}
\note{
This code was found in the examples of \link[base]{is.integer}.
}
\examples{

is.wholenumber(1.2)
is.wholenumber(1)

}
\keyword{internal}
