% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{outsiders}
\alias{outsiders}
\alias{\%[]\%}
\alias{\%()\%}
\alias{\%[)\%}
\alias{\%(]\%}
\title{Inside interval}
\usage{
x \%[]\% j

x \%()\% j

x \%[)\% j

x \%(]\% j
}
\arguments{
\item{x}{A vector}

\item{j}{A range}
}
\value{
True or false
}
\description{
Decide if a value \code{x} falls within an interval \code{j[1],j[2]}

Closed interval

Open interval

Closed interval (left)

Closed interval (right)
}
\examples{

# Closed interval
5\%[]\%c(1,5)

# Open interval
5\%()\%c(1,5)

# Closed interval left
5\%[)\%c(1,5)

# Closed interval right
5\%(]\%c(1,5)

}
\seealso{
\link{[DescTools]}

\link{[DescTools]}

\link{[DescTools]}

\link{[DescTools]}
}
