% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{extractors}
\alias{extractors}
\alias{\%[f\%}
\alias{\%[l\%}
\alias{\%[\%}
\alias{\%]\%}
\alias{\%f]\%}
\alias{\%l]\%}
\title{Extract vectors by index}
\usage{
x \%[f\% v

x \%[l\% v

x \%[\% i

x \%]\% i

x \%f]\% v

x \%l]\% v
}
\arguments{
\item{x}{A vector}

\item{v}{A value of which the first or last occurrence in \code{x} will be used as an index}

\item{i}{An index or two element vector \code{c(lo,hi)} indicating a range to extract}
}
\value{
A vector extracted from the front, rear, or, range of \code{x}. Either based on an index or the first or last occurrence of a value.
}
\description{
Extract vector front/rear up and untill and index \code{i} or the first/last occurence of \code{v} in \code{x}.
}
\note{
The function provided for symmetry, character lengths of \code{x\%]\%i} and \code{x[1:i]} are equal.
}
\examples{

z <- letters

# Extract front by first occurrence of value
z \%[f\% "n"

# Extract front by index
x <- rnorm(100)
x \%[\% 10

# Extract rear by index
x \%]\% 90

# Extract by indices if a range is provided
x \%]\% c(4,30)
z \%[\% c(6,10)

}
