% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invictor.R
\name{extractors}
\alias{extractors}
\alias{\%[1\%}
\alias{\%[\%}
\title{Extract vectors by index}
\usage{
x \%[1\% i

x \%[\% i
}
\arguments{
\item{x}{A vector}

\item{i}{An index}

\item{j}{A two element vector \code{}  indicating a range to extract}
}
\value{
A vector extracted from the front, rear, or, center of \code{x}. Either based on an index or the occurrence of a value.

The contents of vector \code{x} in the range \code{x[1:i]}
}
\description{
Extract vectors by index

Extract vector front up and untill first occurence of \code{i} in \code{x}

Extract vector front by index
}
\note{
This is provided for symmetry, character lengths of \code{x\%]\%i} and \code{x[1:i]} are equal
}
\examples{

# Extract front by value
z <- letters

z\%[1\%"n"

# Extract front by index
x <- rnorm(100)
x\%[\%10

# Extract rear by index
90\%]\%x

# Extract center by indices
x\%[.]\%c(4,30)
z\%[.]\%c(6,10)

}
