% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invctr-package.R
\docType{package}
\name{invctr-package}
\alias{invctr}
\alias{invctr-package}
\title{invctr: Infix Functions For Vector Operations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Vector operations between grapes: An infix-only package! The 'invctr' functions perform common and less common operations on vectors, data frames matrices and list objects: - Extracting a value (range), or, finding the indices of a value (range). - Trimming, or padding a vector with a value of your choice. - Simple polynomial regression. - Set and membership operations. - General check & replace function for NAs, Inf and other values.
}
\examples{
cat("OK")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FredHasselman/invctr}
  \item Report bugs at \url{https://github.com/FredHasselman/invctr/issues}
}

}
\author{
\strong{Maintainer}: Fred Hasselman \email{fred.hasselman@ru.nl} (\href{https://orcid.org/0000-0003-1384-8361}{ORCID})

}
\keyword{internal}
